backend/requirements.txt
# Dependencias del proyecto
Django==5.2.7
djangorestframework==3.15.2
paho-mqtt==2.1.0
django-cors-headers==4.5.0
python-dotenv==1.0.1
django-extensions==3.2.3
matplotlib==3.9.2
plotly==5.24.1

# Para desarrollo
ipython==8.29.0

# Para producción (opcional)
gunicorn==23.0.0
psycopg2-binary==2.9.10